Deleted unused data files, separated the data model for NetflixShow into a separate model-file, added RegEx-query for titles and sorting alphabetically, updated the 'first page' with endpoint info
